exchange model is responsible for updating account balances.  if this fails, only destroy the associated req if it was just created as a non-active req

+++ app/models/exchange.rb
@@ -65,7 +65,9 @@ class Exchange < ActiveRecord::Base
         end
       end
     rescue
-      self.req.destroy
+      unless self.req.active?
+        self.req.destroy
+      end
     end
   end
